Remove redundant as u32 casts

+++ examples/rectangle.rs
@@ -19,7 +19,7 @@ fn redraw(buffer: &mut Buffer<'_, impl HasDisplayHandle, impl HasWindowHandle>, 
                 let red = (x & 0xff) ^ (y & 0xff);
                 let green = (x & 0x7f) ^ (y & 0x7f);
                 let blue = (x & 0x3f) ^ (y & 0x3f);
-                (blue | (green << 8) | (red << 16)) as u32
+                blue | (green << 8) | (red << 16)
             };
             buffer[(y * width + x) as usize] = value;
         }
